refacto(oktaApiAuth): split oktaApiAuth package into multiple files
