+++ pkg/oktaApiAuth/api.go
@@ -1,0 +1,194 @@
+package oktaApiAuth
+
+import (
+	"bytes"
+	"crypto/sha256"
+	"crypto/tls"
+	"crypto/x509"
+	"encoding/base64"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"net/http"
+	"net/url"
+	"slices"
+	"strings"
+	"time"
+
+	log "github.com/sirupsen/logrus"
+)
+
+// Prepare an http client with a safe TLS config
+// validate the server public key against our list of pinned key fingerprint
+func (auth *OktaApiAuth) InitPool() error {
+	if rawURL, err := url.Parse(auth.ApiConfig.Url); err != nil {
+		return err
+	} else {
+		var port string
+		if port = rawURL.Port(); port == "" {
+			port = "443"
+		}
+		// Connect to the server, fetch its public key and validate it against the
+		// base64 digest in pinset slice
+		tcpURL := fmt.Sprintf("%s:%s", rawURL.Hostname(), port)
+		conn, err := tls.Dial("tcp", tcpURL, &tls.Config{InsecureSkipVerify: true})
+		if err != nil {
+			log.Errorf("Error in Dial: %s", err)
+			return err
+		}
+		defer conn.Close()
+		certs := conn.ConnectionState().PeerCertificates
+		for _, cert := range certs {
+			if !cert.IsCA {
+				// Compute public key base64 digest
+				derPubKey, err := x509.MarshalPKIXPublicKey(cert.PublicKey)
+				if err != nil {
+					return err
+				}
+				pubKeySha := sha256.Sum256(derPubKey)
+				digest := base64.StdEncoding.EncodeToString([]byte(string(pubKeySha[:])))
+
+				if !slices.Contains(auth.ApiConfig.AssertPin, digest) {
+					log.Errorf("Refusing to authenticate because host %s failed %s\n%s\n%s",
+						rawURL.Hostname(),
+						"a TLS public key pinning check.",
+						"Update your \"pinset.cfg\" file or ",
+						"contact support@okta.com with this error message")
+					return errors.New("Server pubkey does not match pinned keys")
+				}
+			}
+		}
+	}
+
+	tlsCfg := &tls.Config{
+		InsecureSkipVerify: false,
+		MinVersion:         tls.VersionTLS12,
+		CipherSuites: []uint16{
+			// TLS 1.2 safe cipher suites
+			tls.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
+			tls.TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
+			tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
+			tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
+			tls.TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256,
+			tls.TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256,
+			// TLS 1.3 cipher suites
+			tls.TLS_AES_128_GCM_SHA256,
+			tls.TLS_AES_256_GCM_SHA384,
+			tls.TLS_CHACHA20_POLY1305_SHA256,
+		},
+	}
+	t := &http.Transport{
+		MaxIdleConns:        5,
+		MaxConnsPerHost:     5,
+		MaxIdleConnsPerHost: 5,
+		TLSClientConfig:     tlsCfg,
+	}
+	auth.pool = &http.Client{
+		Timeout:   10 * time.Second,
+		Transport: t,
+	}
+	return nil
+}
+
+func (auth *OktaApiAuth) Pool() *http.Client {
+	return auth.pool
+}
+
+// Do an http request to the Okta API using the path and payload provided
+func (auth *OktaApiAuth) oktaReq(method string, path string, data map[string]string) (a map[string]interface{}, err error) {
+	u, _ := url.ParseRequestURI(auth.ApiConfig.Url)
+	u.Path = fmt.Sprintf("/api/v1%s", path)
+
+	ssws := fmt.Sprintf("SSWS %s", auth.ApiConfig.Token)
+	headers := map[string]string{
+		"User-Agent":    auth.userAgent,
+		"Content-Type":  "application/json",
+		"Accept":        "application/json",
+		"Authorization": ssws,
+	}
+	if auth.UserConfig.ClientIp != "" {
+		headers["X-Forwarded-For"] = auth.UserConfig.ClientIp
+	}
+
+	var r *http.Request
+	var dataReader *bytes.Reader
+	if method == http.MethodPost {
+		jsonData, err := json.Marshal(data)
+		if err != nil {
+			log.Errorf("Error marshaling request payload: %s", err)
+			return nil, err
+		}
+		dataReader = bytes.NewReader(jsonData)
+	} else {
+		dataReader = bytes.NewReader([]byte{})
+	}
+	r, err = http.NewRequest(method, u.String(), dataReader)
+	if err != nil {
+		log.Errorf("Error creating http request: %s", err)
+		return nil, err
+	}
+	for k, v := range headers {
+		r.Header.Add(k, v)
+	}
+	resp, err := auth.pool.Do(r)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+	jsonBody, err := io.ReadAll(resp.Body)
+	if err != nil {
+		log.Errorf("Error reading Okta API response: %s", err)
+		return nil, err
+	}
+	// TODO: return an interface{} and have the "client" functions handle properly
+	// what they expect
+	if strings.HasPrefix(string(jsonBody), "{") {
+		err = json.Unmarshal(jsonBody, &a)
+		if err != nil {
+			log.Errorf("Error unmarshaling Okta API response: %s", err)
+			return nil, err
+		}
+	} else {
+		var res []interface{}
+		err = json.Unmarshal(jsonBody, &res)
+		if err != nil {
+			log.Errorf("Error unmarshaling Okta API response: %s", err)
+			return nil, err
+		}
+		a = make(map[string]interface{})
+		a["data"] = res
+	}
+	return a, nil
+}
+
+// Call the preauth Okta API endpoint
+func (auth *OktaApiAuth) preAuth() (map[string]interface{}, error) {
+	// https://developer.okta.com/docs/reference/api/authn/#primary-authentication-with-public-application
+	data := map[string]string{
+		"username": auth.UserConfig.Username,
+		"password": auth.UserConfig.Password,
+	}
+	return auth.oktaReq(http.MethodPost, "/authn", data)
+}
+
+// Call the MFA auth Okta API endpoint
+func (auth *OktaApiAuth) doAuth(fid string, stateToken string) (map[string]interface{}, error) {
+	// https://developer.okta.com/docs/reference/api/authn/#verify-call-factor
+	path := fmt.Sprintf("/authn/factors/%s/verify", fid)
+	data := map[string]string{
+		"fid":        fid,
+		"stateToken": stateToken,
+		"passCode":   auth.UserConfig.Passcode,
+	}
+	return auth.oktaReq(http.MethodPost, path, data)
+}
+
+// Cancel an authentication transaction
+func (auth *OktaApiAuth) cancelAuth(stateToken string) (map[string]interface{}, error) {
+	// https://developer.okta.com/docs/reference/api/authn/#cancel-transaction
+	data := map[string]string{
+		"stateToken": stateToken,
+	}
+	return auth.oktaReq(http.MethodPost, "/authn/cancel", data)
+}
